NRE при отключении отладки от работающего процесса
