+++ src/ScriptEngine/Machine/MachineStopReason.cs
@@ -1,0 +1,19 @@
+﻿/*----------------------------------------------------------
+This Source Code Form is subject to the terms of the 
+Mozilla Public License, v.2.0. If a copy of the MPL 
+was not distributed with this file, You can obtain one 
+at http://mozilla.org/MPL/2.0/.
+----------------------------------------------------------*/
+
+using System.Threading;
+
+namespace ScriptEngine.Machine
+{
+    public enum MachineStopReason
+    {
+        Breakpoint,
+        BreakpointConditionError,
+        Step,
+        Exception
+    }
+}
